requirements.txt
# Wellnix - AI Health Intelligence Platform
# Core Framework
Flask>=3.0.0
Flask-Cors>=4.0.0
Flask-SQLAlchemy>=3.1.0
Flask-Login>=0.6.3
Flask-Bcrypt>=1.0.1
Werkzeug>=3.0.0
Jinja2>=3.1.0
itsdangerous>=2.1.0
gunicorn>=21.0.0

# Database
SQLAlchemy>=2.0.0
greenlet>=3.0.0

# Authentication & Validation
email-validator>=2.0.0
bcrypt>=4.0.0
authlib>=1.3.0
cryptography>=41.0.0

# HTTP Client
requests>=2.31.0

# Environment
python-dotenv>=1.0.0

# AI/ML - Nutrition
easyocr>=1.7.0
pillow>=10.0.0
numpy>=1.24.0

# AI/ML - Movement Analysis
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
opencv-python>=4.8.0
moviepy>=1.0.3

# JWT Authentication
PyJWT>=2.8.0

# Async Task Queue
celery>=5.3.0
redis>=5.0.0

# Utilities
tqdm>=4.65.0
PyYAML>=6.0.0
python-dateutil>=2.8.0
typing_extensions>=4.5.0
